// HelpDeskFE/HelpDeskFETp/src/api/model/getApiResolutionsParams.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * HelpDesk API
 * OpenAPI spec version: v1
 */

export type GetApiResolutionsParams = {
IdTicket?: number;
IdUser?: number;
IdSolutionStatus?: number;
IdDevice?: number;
IdPriority?: number;
DateFrom?: string;
DateTo?: string;
Keyword?: string;
PageNumber?: number;
PageSize?: number;
};