// HelpDeskFE/HelpDeskFETp/src/api/model/updateResolutionDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * HelpDesk API
 * OpenAPI spec version: v1
 */

export interface UpdateResolutionDto {
  /** @minimum 1 */
  idTicket: number;
  /**
     * @minLength 0
     * @maxLength 500
     */
  actionTaken: string;
  /** @minimum 1 */
  idSolutionStatus: number;
  /**
     * @minLength 0
     * @maxLength 500
     */
  rootCause: string;
  /**
     * @minLength 0
     * @maxLength 500
     * @nullable
     */
  preventiveMeasures?: string | null;
  /**
     * @minLength 0
     * @maxLength 500
     * @nullable
     */
  observation?: string | null;
  /**
     * @minLength 0
     * @maxLength 500
     * @nullable
     */
  secondObservation?: string | null;
  /** @minimum 1 */
  idPriority: number;
  /** @nullable */
  idDevice?: number | null;
  /**
     * @minimum 0.1
     * @maximum 999.99
     */
  solutionTime: number;
}